--- 2.5D Beetle Game/Assets/PlayerMovement.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 70398722f0a4d704aa2f63a33881bd13
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
