+++ Assets/Beetle placeholder.mat.meta
