+++ ProjectSettings/NavMeshAreas.asset
